No-op implementation of Aetos endpoints
This is a "no-op" reverse-proxy implementation. It doesn't
yet do anything around multi-tenancy. This implements all
the API endpoints which are used by the observabilityclient
and by watcher except for the admin endpoints. Admin
endpoints and potentially an extension for the whole Prometheus
API https://prometheus.io/docs/prometheus/latest/querying/api/
will be added in the future (although support for endpoints
used by the observabilityclient, Aodh and Watcher will be
the priority).

Followup changes will include policy handling and query modifications
using the observabilityclient rbac module to achieve multi-tenancy
enforcement for requests comming from non-admin users.

This is missing tests for now as this is mostly just the minimal
skeleton for now. The plan is to start implementing tests together
with proper implementation of each endpoint in followup changes.

Change-Id: Ief18baeee3fcaccc10da0b76c5b50cd70d4a92cf