Move unit tests to functional tests
This moves the current tests into functional tests. The test coverage
remains pretty much the same, the only difference is, that instead
of mocking certain parts of pecan and calling the individual functions
inside Aetos, we now run a test app and execute http requests against it.

The reason for this change is, that I started implementing policy
handling and it seemed pretty difficult to test with the previous
style of tests. We'd basically need to mock a lot of objects,
which are external to Aetos, which would require knowledge of
their insides and it'd be prone to failures when some of the
mocked dependencies change.

This also follows what I've seen in Aodh tests, where the Aodh api
is tested very similarly to this.

Change-Id: Ia3a712cdcdf2d38fdf2dd489b5ce2afd567e7cd9