Add policy and tenancy for query endpoint
This adds the telemetry:query and telemetry:query:all_projects
policies, which are used for the query endpoint. telemetry:query
is accessible by project readers and telemetry:query:all_projects
is accessible by admins and services. This allows to distinguish
between a "low privilege" and "high privilege" access.

When accessing the endpoint with a low privilege, the
PromQLRbac.modify_query from observabilityclient is used to
add project label to each query to enforce tenancy.

When accessing with high privilege, the query is relayed to
prometheus without modification allowing to query metrics across
projects or to query metrics which have no project label.

This completes the work on the query endpoint code. Once this
patch goes through the review and merges, it'll be replicated
in a similar way across all other endpoints, which will complete
the minimal functionality of Aetos.

Change-Id: Ic8de79094f9d2a3b629ef3b59f2c5931421bb3b1