Add tenancy enforcement for all endpoints
This follows a similar template to enforce tenancy for the
rest of the endpoints as is already implemented for the query
endpoint.

To recap, there are 2 modes of access configurable through
policies. Higher privilege access is by default enabled for
admin and service users and allows access to all metrics, while
lower privilege access is by default enabled for readers and
allows access only to metrics from the project supplied in the
token.

Change-Id: I1ff8c7a227a5b5b91ddf3fb9800b7530d3ea542c