Allow match[] parameter in label and labels
This will allow users to restrict the amount of values returned.
For example admins will be able to restrict the labels to just
one project similarly to what's being done automatically when
a request is sent from non-admin user.

Also modify similar endpoints slightly to follow the same pattern
of having *args, **kwargs arguments and checking the length of
args.

Change-Id: I408a5d276a873912f6d0845fd985d44ceda4b95f