Reduce Prometheus requests for modify_query calls
Each call to modify_query will send a request to Prometheus to
determine all metric names saved in Prometheus unless we
provide the list with the modify_query call. We use the modify_query
in two places in Aetos, one is the query controller, the other one
is the process_matches function. In the query controller, we call
modify_query just once, so the code can stay as is. In process_matches
we call modify_query in a loop, so we send a request to Prometheus
to retrieve a list of metrics before the loop and provide the list
to modify_query to to avoid sending a request to Prometheus with each
iteration of the loop.

Change-Id: I5f5c82ce25033748c321dd836256094fbcf305ef
Signed-off-by: Jaromir Wysoglad <jwysogla@redhat.com>