rely on devstack to install uwsgi
installation of uwsgi from pypi is flaky
as the compilation of the uwsgi c module sometimes
fails for non determinisic reasons.

Devstack already will install uwsgi for us and
it will prefer to use the distro package if possible avoiding
the need to compile it in the job. this is both faster and
more reliable reducing rechecks.

Closes-Bug: #2126581
Change-Id: I5048360c698d0c60c1e61f9a1163cfaac3961b6b
Signed-off-by: Sean Mooney <work@seanmooney.info>